enhance ButtonTemplate and writeComponentFile for better file handling and class management

+++ src/writeComponentFile.ts
@@ -1,4 +1,5 @@
 import path from "path";
+import readline from "readline";
 import { generateComponentCode } from "./utils/generateComponentCode";
 import { ensureDirectoryExists, ensureFileDoesNotExist, writeFile } from "./utils/fileUtils";
 import { ComponentType } from "./utils/types";
@@ -23,11 +24,26 @@ export const writeComponentFile = (
         const fileExtension = isTypescript ? "tsx" : "jsx";
         const filePath = path.join(PATHS.COMPONENTS_DIR, `${name}.${fileExtension}`);
 
-        ensureFileDoesNotExist(filePath);
+        if (ensureFileDoesNotExist(filePath)) {
+            writeFile(filePath, componentCode);
+            console.log("\n" + MESSAGES.COMPONENT_GENERATED(name, type, filePath));
+        } else {
+            const rl = readline.createInterface({
+                input: process.stdin,
+                output: process.stdout
+            });
 
-        writeFile(filePath, componentCode);
-
-        console.log("\n" + MESSAGES.COMPONENT_GENERATED(name, type, filePath));
+            rl.question(`The file ${filePath} already exists. \nDo you want to overwrite it? (yes/no): (Default: no) `, (answer) => {
+                const normalizedAnswer = answer.trim().toLowerCase();
+                if (normalizedAnswer === 'y' || normalizedAnswer === 'yes') {
+                    writeFile(filePath, componentCode);
+                    console.log("\n" + MESSAGES.COMPONENT_GENERATED(name, type, filePath));
+                } else {
+                    console.log("\n" + MESSAGES.COMPONENT_NOT_OVERWRITTEN(name, type, filePath));
+                }
+                rl.close();
+            });
+        }
 
         if (isTypescript) {
             console.log("\n" + MESSAGES.TYPESCRIPT_REMINDER);
